+++ src/com/martinleopold/pui/Slider.java
@@ -31,7 +31,7 @@ public class Slider extends Widget {
 	float max = 1;
 	boolean sliding = false;
 			
-	public Slider(PUI pui, float x, float y, float width, float height) {
+	public Slider(PUI pui, int x, int y, int width, int height) {
 		super(pui, x, y, width, height);
 	}
